requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.5
rdflib>=7.0.0
pymantic>=0.1.3
pydantic>=2.5.2
python-dotenv>=1.0.0
jinja2>=3.1.0
PyYAML>=5.4
requests>=2.31.0

# LLM/ChatGPT Integration
openai>=1.3.0

# Semantic Module
networkx>=3.1
pyvis>=0.3.1
jsonschema>=4.0.0

# Development Tools
pytest>=7.0.0
pylint>=2.15.0
